Use HTML::TreeBuilder to find sales_path and financial_path
Looking for login notifications builds a tree for the login response, but if there weren't any notifications the login response is actually the main menu. If the main menu is already parsed into a tree, there's no sense in not using it. Besides, this is probably better than using regex's.