New algorithm for finding the financial reports table
The table's location in the page structure isn't fixed and it seems to have changed again. Instead of using the table's structural address, find it by first finding the "itemsPerPage" input element. Then, look upwards to find the 2nd enclosing table, which should be the table with the list of reports.

This still isn't the most reliable way of finding the reports, but it will have to do until Apple sees fit to id or name the reports table properly.